requirements.txt
streamlit
pandas
matplotlib
seaborn
wordcloud
urlextract
emoji
google-genai
pydantic